+++ client/src/components/Button/Button.stories.tsx
@@ -1,34 +1,40 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import { Button } from './Button'
-import { ConfigProvider } from 'antd'
-import { theme } from '../../theme/override'
+import type { Meta, StoryObj } from '@storybook/react';
+import { Button } from './Button';
+import { ConfigProvider } from 'antd';
+import { theme } from '../../theme/override';
 
 const meta: Meta<typeof Button> = {
   component: Button,
   tags: ['autodocs'],
-  decorators: [(Story) => <ConfigProvider theme={theme}><Story /></ConfigProvider>]
-}
+  decorators: [
+    (Story) => (
+      <ConfigProvider theme={theme}>
+        <Story />
+      </ConfigProvider>
+    ),
+  ],
+};
 
-type Story = StoryObj<typeof Button>
+type Story = StoryObj<typeof Button>;
 
 export const Default: Story = {
   args: {
-    children: 'Button'
-  }
-}
+    children: 'Button',
+  },
+};
 
 export const Loading: Story = {
   args: {
     loading: true,
-    children: 'Button'
-  }
-}
+    children: 'Button',
+  },
+};
 
 export const Disabled: Story = {
   args: {
     disabled: true,
-    children: 'Button'
-  }
-}
+    children: 'Button',
+  },
+};
 
-export default meta
+export default meta;