+++ client/src/components/Footer/footer.styled.ts
@@ -1,11 +1,11 @@
-'use client'
+'use client';
 
-import styled, { css } from "styled-components"
-import { LinkProps, TextProps } from "../Typography"
-import Typography from '../Typography/typography.styled'
-import { colorMuted } from "../../theme"
+import styled, { css } from 'styled-components';
+import { LinkProps, TextProps } from '../Typography';
+import Typography from '../Typography/typography.styled';
+import { colorMuted } from '../../theme';
 
-const {Link, Text} = Typography
+const { Link, Text } = Typography;
 
 const SocialLink = styled(Link)<LinkProps>`
   font-size: 20px;
@@ -13,20 +13,22 @@ const SocialLink = styled(Link)<LinkProps>`
   * {
     color: ${colorMuted};
   }
-`
+`;
 
 const Copyright = styled(Text)<TextProps>`
-  ${props => props.$lead && css`
-    font-size: 8px;
+  ${(props) =>
+    props.$lead &&
+    css`
+      font-size: 8px;
 
-    && {
-      * {
-        font-size: 8px;
+      && {
+        * {
+          font-size: 8px;
+        }
       }
-    }
-  `}
-`
+    `}
+`;
 
-const Styled = {SocialLink, Copyright}
+const Styled = { SocialLink, Copyright };
 
-export default Styled
+export default Styled;