+++ client/src/components/Header/Header.stories.tsx
@@ -1,18 +1,24 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import { ConfigProvider } from 'antd'
-import { Header } from './Header'
-import { theme } from '../../theme/override'
+import type { Meta, StoryObj } from '@storybook/react';
+import { ConfigProvider } from 'antd';
+import { Header } from './Header';
+import { theme } from '../../theme/override';
 
 const meta: Meta<typeof Header> = {
   component: Header,
   tags: ['autodocs'],
-  decorators: [(Story) => <ConfigProvider theme={theme}><Story /></ConfigProvider>],
-}
+  decorators: [
+    (Story) => (
+      <ConfigProvider theme={theme}>
+        <Story />
+      </ConfigProvider>
+    ),
+  ],
+};
 
-type Story = StoryObj<typeof Header>
+type Story = StoryObj<typeof Header>;
 
 export const Default: Story = {
-  render: () => <Header />
-}
+  render: () => <Header />,
+};
 
-export default meta
+export default meta;