+++ client/src/components/Pagination/pagination.styled.ts
@@ -1,8 +1,8 @@
-'use client'
+'use client';
 
-import { breakpoints } from "../../shared/constants"
-import { Space, SpaceProps } from "antd"
-import styled from "styled-components"
+import { breakpoints } from '../../shared/constants';
+import { Space, SpaceProps } from 'antd';
+import styled from 'styled-components';
 
 const Pagination = styled(Space)<SpaceProps>`
   padding: 0 24px;
@@ -11,8 +11,8 @@ const Pagination = styled(Space)<SpaceProps>`
   @media screen and (min-width: ${breakpoints.lg}) {
     padding: 0;
   }
-`
+`;
 
-const Styled = {Pagination}
+const Styled = { Pagination };
 
-export default Styled
+export default Styled;