+++ client/src/components/Typography/Title.stories.tsx
@@ -1,21 +1,27 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import { ConfigProvider } from 'antd'
-import { Title } from './Title'
-import { theme } from '../../theme/override'
+import type { Meta, StoryObj } from '@storybook/react';
+import { ConfigProvider } from 'antd';
+import { Title } from './Title';
+import { theme } from '../../theme/override';
 
 const meta: Meta<typeof Title> = {
   title: 'Typography/Title',
   component: Title,
   tags: ['autodocs'],
-  decorators: [(Story) => <ConfigProvider theme={theme}><Story /></ConfigProvider>]
-}
+  decorators: [
+    (Story) => (
+      <ConfigProvider theme={theme}>
+        <Story />
+      </ConfigProvider>
+    ),
+  ],
+};
 
-type Story = StoryObj<typeof Title>
+type Story = StoryObj<typeof Title>;
 
 export const Heading2: Story = {
   args: {
-    children: 'ULTRICIES SED MAGNA EUISMOD ENIM VITAE GRAVIDA'
+    children: 'ULTRICIES SED MAGNA EUISMOD ENIM VITAE GRAVIDA',
   },
-}
+};
 
-export default meta
+export default meta;