+++ client/src/components/Typography/interface.ts
@@ -1,14 +1,14 @@
-import { TextProps as AntTextProps } from 'antd/es/typography/Text'
-import { LinkProps as NextLinkProps } from 'next/link'
+import { TextProps as AntTextProps } from 'antd/es/typography/Text';
+import { LinkProps as NextLinkProps } from 'next/link';
 
 type TextPropsExtended = {
-  $lead?: boolean
-}
+  $lead?: boolean;
+};
 
 type LinkPropsExtended = {
-  $secondary?: boolean
-  $heading?: boolean
-}
+  $secondary?: boolean;
+  $heading?: boolean;
+};
 
-export type TextProps = AntTextProps & TextPropsExtended
-export type LinkProps = NextLinkProps & LinkPropsExtended
+export type TextProps = AntTextProps & TextPropsExtended;
+export type LinkProps = NextLinkProps & LinkPropsExtended;