+++ client/src/components/navbar/Navbar.stories.tsx
@@ -1,21 +1,27 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import { ConfigProvider } from 'antd'
-import { Navbar } from './Navbar'
-import { theme } from '../../theme/override'
+import type { Meta, StoryObj } from '@storybook/react';
+import { ConfigProvider } from 'antd';
+import { Navbar } from './Navbar';
+import { theme } from '../../theme/override';
 
 const meta: Meta<typeof Navbar> = {
   component: Navbar,
   tags: ['autodocs'],
-  decorators: [(Story) => <ConfigProvider theme={theme}><Story /></ConfigProvider>],
+  decorators: [
+    (Story) => (
+      <ConfigProvider theme={theme}>
+        <Story />
+      </ConfigProvider>
+    ),
+  ],
   parameters: {
-    layout: 'fullscreen'
-  }
-}
+    layout: 'fullscreen',
+  },
+};
 
-type Story = StoryObj<typeof Navbar>
+type Story = StoryObj<typeof Navbar>;
 
 export const Default: Story = {
-  render: () => <Navbar />
-}
+  render: () => <Navbar />,
+};
 
-export default meta
+export default meta;